commit 868e1ac07011f1f9626a73a411c31ebbdd462e3d
    12554 want sdev plugin framework (fix header issue)
commit b237158d576c3f39f35d97c4dd214c07273ddde4
    12674 want mac rx barrier function
commit b939b187ee16d21b506005807c652e2c7bdfe243
    12645 vds: smatch and NULL pointer errors
commit f8927fa6636832af96f657f29571bdbd213c6247
    12554 want sdev plugin framework
commit 725953abcc30bcf6f15d7b8a313ab0749275d95f
    12610 rmcadm: NULL pointer errors
commit cb453c7ae4b42ac7a2d7ce155181572a96d69b33
    12650 ds: smatch and NULL pointer errors
commit 29368e865707707ba77989693d34358ebfa08757
    12644 vdc: NULL pointer errors
commit 59b05390e07c3b0e69288f9a732dc5b90d464721
    12539 mdb: NULL pointer errors
commit 8124b8119d8ab7f80f1bd8b094cd96356f531b54
    12597 want libutempter
commit c8907381b10259bed3a871a2881c0f3b7287ab3c
    12705 remove __builtin_alloca() declaration from string.h

	Conflicts:
	usr/src/uts/common/sys/mac_impl.h
	usr/src/uts/common/sys/Makefile

+++ usr/src/cmd/mdb/sun4u/modules/unix/unix.c
@@ -676,7 +676,7 @@ ttrace_walk_init(mdb_walk_state_t *wsp)
 	ttrace_cpu_data_t *tc;
 	struct trap_trace_record *buf;
 
-	if (wsp->walk_addr != NULL) {
+	if (wsp->walk_addr != (uintptr_t)NULL) {
 		mdb_warn("ttrace only supports global walks\n");
 		return (WALK_ERR);
 	}
@@ -965,7 +965,7 @@ httrace_walk_init(mdb_walk_state_t *wsp)
 	struct htrap_trace_record *buf;
 	htrap_trace_hdr_t *hdr;
 
-	if (wsp->walk_addr != NULL) {
+	if (wsp->walk_addr != (uintptr_t)NULL) {
 		mdb_warn("httrace only supports global walks\n");
 		return (WALK_ERR);
 	}
@@ -1347,7 +1347,7 @@ vecint_walk_init(mdb_walk_state_t *wsp)
 {
 	vecint_walk_data_t	*vecint;
 
-	if (wsp->walk_addr != NULL) {
+	if (wsp->walk_addr != (uintptr_t)NULL) {
 		mdb_warn("vecint walk only supports global walks\n");
 		return (WALK_ERR);
 	}
@@ -1378,13 +1378,14 @@ vecint_walk_step(mdb_walk_state_t *wsp)
 	intr_vec_t		iv;
 	int			status;
 
-	if (wsp->walk_addr == NULL) {
+	if (wsp->walk_addr == (uintptr_t)NULL) {
 		while ((vecint->vec_idx < max) && ((wsp->walk_addr =
-		    (uintptr_t)vecint->vec_table[vecint->vec_idx++]) == NULL))
+		    (uintptr_t)vecint->vec_table[vecint->vec_idx++]) ==
+		    (uintptr_t)NULL))
 			continue;
 	}
 
-	if (wsp->walk_addr == NULL)
+	if (wsp->walk_addr == (uintptr_t)NULL)
 		return (WALK_DONE);
 
 	status = wsp->walk_callback(wsp->walk_addr, wsp->walk_data,
@@ -1443,7 +1444,7 @@ softint_walk_init(mdb_walk_state_t *wsp)
 {
 	intr_vec_t	*list;
 
-	if (wsp->walk_addr != NULL) {
+	if (wsp->walk_addr != (uintptr_t)NULL) {
 		mdb_warn("softint walk only supports global walks\n");
 		return (WALK_ERR);
 	}
@@ -1471,7 +1472,7 @@ softint_walk_step(mdb_walk_state_t *wsp)
 	intr_vec_t		iv;
 	int			status;
 
-	if (wsp->walk_addr == NULL)
+	if (wsp->walk_addr == (uintptr_t)NULL)
 		return (WALK_DONE);
 
 	status = wsp->walk_callback(wsp->walk_addr, wsp->walk_data,
@@ -1568,7 +1569,7 @@ mutex_owner_step(mdb_walk_state_t *wsp)
 	if (!MUTEX_TYPE_ADAPTIVE(&mtx))
 		return (WALK_DONE);
 
-	if ((owner = (uintptr_t)MUTEX_OWNER(&mtx)) == NULL)
+	if ((owner = (uintptr_t)MUTEX_OWNER(&mtx)) == (uintptr_t)NULL)
 		return (WALK_DONE);
 
 	if (mdb_vread(&thr, sizeof (thr), owner) != -1)
